requirements.txt
openai
sounddevice
scipy
pygame
python-dotenv